chore: rename instance.ts
