--- src/apis/project/deleteProject/axios.ts
@@ -1,5 +1,0 @@
-import instance from '@/apis/axios';
-
-export const deleteProject = async (projectKey: string) => {
-	await instance.delete(`/projects/${projectKey}`);
-};
